backend/requirements.txt
Flask
flask-cors
pandas
python-docx
google-generativeai